requirements.txt

# LOCAL PACKAGES
# -e .

# AWS
boto3==1.34.81

# LOGGING
loguru==0.7.2

# DOCUMENTATION
# mkdocs==1.6.0

# API
openai==1.42.0
openpyxl==3.1.2
pandas==2.2.2
# peft==0.11.1
# pipreqs==0.5.0

# LINTERS
pre-commit==3.7.0
python-docx==1.1.2

# DATABASE
# psycopg2-binary==2.9.9

# UTILITIES
python-dotenv==1.0.1
# requests==2.32.3

# WEBAPP
streamlit==1.38.0
tqdm==4.66.4
typer==0.12.3
